requirements.txt
Django==4.0.2
psycopg2
django-phonenumber-field[phonenumberslite]
djangorestframework
# python manage.py makemigrations GRABIRAPP 